+++ CHANGELOG.rst
@@ -19,6 +19,8 @@ Unreleased
 * Improved tools.vmi.Distributions performance on Windows (PR #270).
 * More corrections to the GUI example: working without the "bases" directory,
   loading "from transform", interface enhancements (PR #277).
+* Improved documentation (PR #283, PR #288).
+* Correctly use quadrants in abel.Transform (PR #287).
 
 v0.8.3 (2019-08-16)
 -------------------